Moved core components to Kotlin only module
