+++ .gitignore
@@ -5,6 +5,6 @@
 /.idea/workspace.xml
 /.idea/libraries
 .DS_Store
-/build
+/**/build/*
 /captures
 .externalNativeBuild
