+++ library-core/src/main/kotlin/com/pusher/platform/Cancelable.kt
