+++ library-core/src/main/kotlin/com/pusher/platform/MediaTypeResolver.kt
@@ -1,0 +1,7 @@
+package com.pusher.platform
+
+import java.io.File
+
+interface MediaTypeResolver {
+    fun fileMediaType(file: File) : String
+}
