+++ library-core/src/main/kotlin/com/pusher/platform/Scheduler.kt
@@ -1,0 +1,15 @@
+package com.pusher.platform
+
+interface MainThreadScheduler : Scheduler
+
+interface Scheduler {
+
+    fun schedule(action: () -> Unit) : ScheduledJob
+    fun schedule(delay: Long, action: () -> Unit) : ScheduledJob
+
+}
+
+interface ScheduledJob {
+    fun cancel()
+}
+
