+++ library-core/src/main/kotlin/com/pusher/platform/network/ConnectivityHelper.kt
@@ -1,0 +1,7 @@
+package com.pusher.platform.network
+
+interface ConnectivityHelper {
+    fun isConnected(): Boolean
+    fun onConnected(retryNow: () -> Unit)
+    fun cancel()
+}
