+++ library-core/src/main/kotlin/com/pusher/platform/subscription/SubscriptionState.kt
