+++ library-core/src/main/kotlin/elements/Headers.kt
@@ -1,0 +1,3 @@
+package elements
+
+typealias Headers = MutableMap<String, List<String>>
