+++ library-core/src/main/kotlin/elements/Subscription.kt
@@ -1,0 +1,5 @@
+package elements
+
+interface Subscription {
+    fun unsubscribe()
+}
