+++ library/build.gradle
@@ -30,12 +30,12 @@ android {
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar', '*.aar'])
     implementation 'com.android.support:appcompat-v7:26.1.0'
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
+    api project(":library-core")
     api 'com.squareup.okhttp3:okhttp:3.9.1'
-    api'com.google.code.gson:gson:2.8.2'
+    api 'com.google.code.gson:gson:2.8.2'
 
     testImplementation 'junit:junit:4.12'
     testImplementation 'com.nhaarman:mockito-kotlin-kt1.1:1.5.0'
