+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':sample_kotlin', ':library', ':sample_java', ':sample_java'
+include ':sample_kotlin', ':library', ':sample_java', ':sample_java', ':library-core'
